rename messages2 package to v20
