+++ ocpp-json/src/main/scala/com/thenewmotion/ocpp/json/Ocpp2Procedure.scala
@@ -3,7 +3,7 @@ package json
 
 import scala.language.higherKinds
 import scala.concurrent.{ExecutionContext, Future}
-import messages2._
+import v20._
 import CsmsReqRes._
 import org.json4s.{DefaultFormats, Extraction, Formats, JValue}
